freedreno: Always use aux-ctx for export blits
If we need to blit a resource to a buffer that can be exported (ie.
because it was suballocated, etc), always use the aux-context.  This
can be called off of the driver thread, so it isn't thread-safe to
use the main context, even if it is not NULL.

Signed-off-by: Rob Clark <rob.clark@oss.qualcomm.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37696>