nvk: Do not use SET_L1_CONFIGURATION on 3D state init
There is no reasons to cut 48KiB of memory out of L1 cache on gfx
considering that we do not have shared memory and that local
memory does not need to be directly addressable.

This is not set by NVIDIA proprietary driver and the golden ctx setup keep it
uninitialized.

Unsure if that will change anything in term of performance but it might reduce
L1 cache usage on 3D.

Signed-off-by: Mary Guillemard <mary@mary.zone>
Reviewed-by: Mel Henning <mhenning@darkrefraction.com>
Tested-by: Thomas H.P. Andersen <phomes@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/40753>