aux/trace: remove -I argument
enums2names.py is only uses in one place. I propose to remove the -I
argument that is not strictly necessary as we can already get the
header name from the `-H` argument.

That modification is motivated by the need to help ninja-to-soong to
generate proper rule for the Android build system.

ninja-to-soong can't differenciate output file location and a string
matching the output file name.

Ref #14072

Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37785>