glapi: static_data: do not use __file__ to get gl symbols file
Use an explicit path for libgl-symbols.txt from the build system
instead of reconstructing it from __file__.

The issue is that for Android build system, everything is sandboxed
and that file is not in the same root as the python script. Thus we
need a proper explicit path in meson to be able to translate it to a
legal Android construct that is capable of finding that file.

Update everything using libgl_public_functions to propagate that path.

Ref #14072

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37866>