panvk/v9+: Reduce maxBoundDescriptorSets to 7
Until now, the driver has been using a single set for internal state
while leaving maxBoundDescriptorSets to the remaining 15.

This gives us no room for optimizations of driver sets, which might
become an issue in the future.

To remedy this, we therefore reduce maxBoundDescriptorSets to 7. This
aligns with the proprietary driver and gives us the space to optimize
the driver sets.

We might increase this in the future if we see that we don't need all
the driver sets we now reserve.

Reviewed-by: John Anthony <john.anthony@arm.com>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>
Reviewed-by: Olivia Lee <olivia.lee@collabora.com>
Reviewed-by: Boris Brezillon <boris.brezillon@collabora.com>
Acked-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37978>