nak/cmat: add optimisation to cmat load/store to do 32-bit load for f16vec2
Initial idea and code from Dave, but this is a complete rewrite of the
patch.

The Matrix layouts contain groups of values, for int8 we have vec4 groups,
for fp16, fp32 and int32 we have vec2s. With this we load and store them
as vectors getting rid of a bunch of address calculation.

Reviewed-by: Mel Henning <mhenning@darkrefraction.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37998>