lavapipe: add support for VK_KHR_cooperative_matrix.
This adds support for cooperative matrix to lavapipe.

It uses 8x8 matricies as the size, and loads the first row
of 8 values into a the subgroup.

It stores the B matrix transposed to make the matmul operation
a lot more subgroup friendly.

Reviewed-by: Georg Lehmann <dadschoorse@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38935>