intel/isl: Omit unused space when sizing mipmaps
Some mipmapped surfaces contain unused space in the bottom right side of
the mipmap. Omit as many unused tiles from that area as possible when
computing the size of the surface.

When running a Final Fantasy 14 DX11 trace on DG2, this saves 2MB on a
256x256x32 (3D) B8G8R8A8_UNORM image with 9 mip levels.

Note that in addition to reducing memory usage, this change has a side
effect of affecting tiling decisions. For example, when running the
Total War Warhammer 3 trace on ICL, INTEL_DEBUG=isl emits the following
information for a 4K x 4K, 3-level, R8G8B8A8_UNORM mipmap:

* Omitted 64 4KB page(s) of padding. [...] tiling_flags=+Y0
* Omitted 64 4KB page(s) of padding. [...] tiling_flags=+icl-Yf
* Omitted 256 4KB page(s) of padding. [...] tiling_flags=+icl-Ys
* Saved 192 4KB page(s). [...] tiling_flags=+icl-Ys

This demonstrates the two impacts:

1) ISL created three surfaces, and ultimately chose the Ys-tiled one due
   to the removal of padding resulting in a smaller image size. Without
   the removal, the Ys surface would have been the same size as the Y0
   surface. With the removal, it is 768KB smaller.

2) The surface returned by ISL is 1MB smaller as a result of removing of
   the padding.

Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/39593>