zink/ci: document a nightly failure
This started failing at some point between adeec19aab0 and 2ab63ffdadc,
but the nightly CI was out for a long time in between, so it's hard to
tell what commit it to blame without bisecting. I'm going to leave that
up to someone else for now.

To complicate things further, it seems like this one only fails in the
nightly runs, but not in pre-merge. So we list it as a flake instead of
a consisten failure.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38279>