llvmpipe: elliptical derivative transform for anisotropic filtering
This improves a lot the quality of anisotropic filtering on surfaces
with a high angle of incidence.

This can also be used for non-anisotropic filtering, but the effect isn't
as pronounced, and might not be worth the cost. In fact in my testing,
it didn't seem to be used on Apple hardware.

Based on this excellent article:
https://pema.dev/2025/05/09/mipmaps-too-much-detail/

Acked-by: Roland Scheidegger <roland.scheidegger@broadcom.com>
Acked-by: Konstantin Seurer <konstantin.seurer@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38316>