panvk: Improve base instance tracking for indirect draws
There's an issue with indirect draws where we do descriptor patching on
the command stream if the indirect baseInstance is non-zero.  In order
for this to work, we need to ensure we have N fresh attribute tables and
N fresh resource tables to reference them.  We tried to handle this by
smashing MESA_VK_DYNAMIC_VI to dirty whenever we might hit this case.
However, we originally placed this check before se actually computed
vi.attribs_changing_on_base_instance so the check was wrong if the
previous VI set didn't have VK_VERTEX_INPUT_RATE_INSTANCE anywhere.

Fortunately, if we ever hit that case and it mattered, VI would be dirty
so would re-emit anyway.

This commit adds a new BASE_INSTANCE driver dirty flag and sets that
whenever the base instance changes or we're an indirect draw.  Then we
can add some helpers which do some sanity assertion checks and make the
correctness far more clear overall.  This also lets us reduce some of
the calculation duplication to prevent things from getting out-of sync
in the future.

Signed-off-by: Olivia Lee <olivia.lee@collabora.com>
Reviewed-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Christian Gmeiner <cgmeiner@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/41654>