panvk/csf: Break VS descriptor setup into a new helper
This and the previous commit are the beginning of an overall restructure
of the state flushing flow.  The idea is that the very first thing
flushed are descriptors, including vertex and fragment attributes.
Importantly, flushing descriptors does not depend on which vertex shader
variant we end up picking in the end.  Both the HW and SW VS will use
the same descriptors.

Signed-off-by: Olivia Lee <olivia.lee@collabora.com>
Reviewed-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Christian Gmeiner <cgmeiner@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/41654>