panvk/jm: Plumb IA state through draw_info
With geometry shaders, we're going to be changing the IA state as we
handle the geometry pipeline and it'll make things much cleaner if we
make it part of the draw info.  To keep things consistent, we refactor
both halves of the driver.  JM is simpler so we start there.

Signed-off-by: Olivia Lee <olivia.lee@collabora.com>
Reviewed-by: Eric R. Smith <eric.smith@collabora.com>
Reviewed-by: Christian Gmeiner <cgmeiner@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/41654>