pan/bi: Run nir_lower_all_phis_to_scalar() late
We were running this in the preprocess step and then trusting that it
would clean up everything before we got to the back-end.  However, we
were running the entire optimization loop in between as well as drivers
potentially adding stuff (since panvk has it's own passes after
postprocess).  Instead, this should be one of the last things run, right
before we go into the back-end.

Cc: mesa-stable
Reviewed-by: Christoph Pillmayer <christoph.pillmayer@arm.com>
Acked-by: Eric R. Smith <eric.smith@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38821>