intel/decoder: make libvulkan_intel to depend on stub decoder when buildtyle=release.
The libvulkan_intel does not need the decoder when buildtype=release
where the debugging is disabled.

However, the decoder implementation is decided by the dep_expat
which may be turned on by like -Dtools=intel and the binary size
of libvulkan_intel increase unexpectedly.

This change creates the stub dependency and decide the exact
decoder dependency of libvulkan_intel by the buildtype.

Test: meson setup builddir -D build-tests=true -Dbuildtype=release --reconfigure && ninja -C builddir && cd builddir && meson test

Reviewed-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Signed-off-by: Andy Hsu <hwandy@google.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38569>