pan/bi: Use the pan_varying_layout for Vallhall+ direct varing load/store
For the descriptor paths, we just assert that the locations match what
we expect and move on.  For the direct paths, we use the offsets in the
layout.  The compiler also now takes an optional layout which the driver
can provide itself if desired.

Reviewed-by: Lorenzo Rossi <lorenzo.rossi@collabora.com>
Acked-by: Eric R. Smith <eric.smith@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38681>