venus: disable TLS ring prio forwarding on Windows
The prios aren't consistent when guest and host are different OS'es.
Even if they are the same, sometimes the forwarding can hit other
permission issues. The entire ring prio thing has to be redesigned, and
before that we just disable it on Windows.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/38969>