ci,crnm: refresh_wait_job as argument
The polling period of the monitoring loop is a constant hardcoded and defined
to an arbitrary default number. This is related to the output reporting to the
user the progress in the pipeline running, so it shows this tool is alive when
it can be generating a huge number of output lines. Transforming it to a
definable variable allows the possibility to experiment with different values,
because of the balance between when a job can be triggered and the verbose
output produced.

Signed-off-by: Sergi Blanch Torne <sergi.blanch.torne@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/39105>